LPS-77357 Move to internal
